+++ Tetris/Tetris/Logic.cc
@@ -25,8 +25,8 @@ Logic::TetrisTable
 Logic::getTable()
 {
 	TetrisTable fullTable = landedTable_;
-	for (int i = 0; i < currentShape_->shape.size(); i++) {
-		for (int j = 0; j < currentShape_->shape.at(i).size(); j++) {
+	for (size_t i = 0; i < currentShape_->shape.size(); i++) {
+		for (size_t j = 0; j < currentShape_->shape.at(i).size(); j++) {
 			fullTable.at(j + currentShape_->topLeft.first).at(i + currentShape_->topLeft.second)
 				= currentShape_->shape.at(i).at(j);
 		}
